# Optimize the ` ToHexString`