--- a/src/Neo.Cryptography.BLS12_381/Scalar.cs
+++ b/src/Neo.Cryptography.BLS12_381/Scalar.cs
@@ -14,6 +14,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Security.Cryptography;
+using System.Text;
 using static Neo.Cryptography.BLS12_381.ConstantTimeUtility;
 using static Neo.Cryptography.BLS12_381.MathUtility;
 using static Neo.Cryptography.BLS12_381.ScalarConstants;
@@ -119,7 +120,10 @@ public readonly struct Scalar : IEquatable<Scalar>, INumber<Scalar>
 
     public override string ToString()
     {
-        return "0x" + ToArray().ToHexString(true);
+        StringBuilder sb = new();
+        sb.Append("0x");
+        sb.Append(ToArray().ToHexString(true));
+        return sb.ToString();
     }
 
     public static bool operator ==(in Scalar a, in Scalar b)