# Description

I tested some RPC interfaces and found that GC consumes a quite bit of CPU.
```
•Runtime Version: V 8.0.824.36612 (built on 17/7/2024 02:12:37)
•CLR Startup Flags: 8388611
•Total CPU Time: 34,098 msec
•Total GC CPU Time: 6,984 msec
•Total Allocs : 8,984.575 MB
•Number of Heaps: 1
•GC CPU MSec/MB Alloc : 0.777 MSec/MB
•Total GC Pause: 7,440.0 msec
•% Time paused for Garbage Collection: 27.6%
•% CPU Time spent Garbage Collecting: 20.5%
•Max GC Heap Size: 194.878 MB
•Peak Process Working Set: 311.857 MB
•Peak Virtual Memory Usage: 2,481,154.531 MB
```

And I found some memory allocations that can be optimized.

For example:
Th current impl of `ToHexString`:
* `StringBuilder` can be pre-allocated.
* `AppendFormat` will box the byte `b` and causes a/some memory allocation(s).
```
        StringBuilder sb = new();
        foreach (var b in bytes)
            sb.AppendFormat("{0:x2}", b);
        return sb.ToString();
```

Perf tests on 256B bytes:
```
BenchmarkDotNet v0.13.12, macOS Ventura 13.7 (22H123) [Darwin 22.6.0]
Intel Core i7-8750H CPU 2.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET SDK 8.0.203
  [Host]     : .NET 8.0.3 (8.0.324.11423), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.3 (8.0.324.11423), X64 RyuJIT AVX2


| Method                   | Mean        | Error     | StdDev    |
|------------------------- |------------:|----------:|----------:|
| ToHex_Optimized          |    339.0 ns |   6.70 ns |   7.71 ns |
| ToHex_Reversed_Optimized |    525.0 ns |  10.44 ns |   9.76 ns |
| ToHexSpan_Optimized      |    880.8 ns |  17.47 ns |  26.69 ns |
| ToHex_Legacy             | 10,236.3 ns | 172.80 ns | 161.63 ns |
| ToHex_Reversed_Legacy    | 10,445.8 ns | 185.36 ns | 206.03 ns |
```


## Type of change

<!-- Please delete options that are not relevant. -->

- [x] Optimization (the change is only an optimization)
- [ ] Style (the change is only a code style for better maintenance or standard purpose)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules