@nan01ab 2024-11-06
> I prefer char array, it will be faster, but if all @neo-project/core is ok, i'm ok

I tested again.
`const string` is faster when `ToHexString([]byte)`, `char[]` is  faster when `ToHexString(reverse = true)`, no difference when `ToHexString(Span)`;

`ToHexString([]byte)` is used in most cases , so  `cost string` is used.