@shargon 2024-11-07
> > I prefer char array, it will be faster, but if all @neo-project/core is ok, i'm ok
> 
> I tested again. `const string` is faster when `ToHexString([]byte)`, `char[]` is faster when `ToHexString(reverse = true)`, no difference when `ToHexString(Span)`;
> 
> `ToHexString([]byte)` is used in most cases , so `cost string` is used(May more friendly for some compiler optimazation)

this always will be slower than direct array access, but it's ok, just a ms

https://github.com/dotnet/runtime/blob/main/src/libraries/System.Private.CoreLib/src/System/String.cs#L750-L753